+++ app/src/main/java/woowacourse/shopping/data/datasource/CartItemDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.datasource
 
 import woowacourse.shopping.data.model.request.CartItemRequest
-import woowacourse.shopping.data.model.response.CartItemResponse
+import woowacourse.shopping.data.model.response.cartitem.CartItemResponse
 import woowacourse.shopping.data.model.response.Quantity
 import woowacourse.shopping.data.service.CartItemService
