+++ app/src/main/java/woowacourse/shopping/data/model/response/cartitem/CartItemContent.kt
@@ -1,7 +1,8 @@
-package woowacourse.shopping.data.model.response
+package woowacourse.shopping.data.model.response.cartitem
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import woowacourse.shopping.data.model.response.product.ProductContent
 
 @Serializable
 data class CartItemContent(
