+++ app/src/main/java/woowacourse/shopping/data/model/response/cartitem/CartItemResponse.kt
@@ -1,11 +1,11 @@
-package woowacourse.shopping.data.model.response
+package woowacourse.shopping.data.model.response.cartitem
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductsResponse(
-    @SerialName("content") val productContent: List<ProductContent>,
+data class CartItemResponse(
+    @SerialName("content") val content: List<CartItemContent>,
     @SerialName("first") val first: Boolean,
     @SerialName("last") val last: Boolean,
 )
